+++ src/Remotes.js
@@ -2,7 +2,8 @@ define(function (require) {
     "use strict";
 
     // Brackets modules
-    var DefaultDialogs  = brackets.getModule("widgets/DefaultDialogs"),
+    var _               = brackets.getModule("thirdparty/lodash"),
+        DefaultDialogs  = brackets.getModule("widgets/DefaultDialogs"),
         Dialogs         = brackets.getModule("widgets/Dialogs"),
         StringUtils     = brackets.getModule("utils/StringUtils");
 
@@ -35,9 +36,21 @@ define(function (require) {
 
     // Implementation
 
-    function getDefaultRemote() {
-        var defaultRemotes = Preferences.get("defaultRemotes") || {};
-        return defaultRemotes[Utils.getProjectRoot()] || "origin";
+    function getDefaultRemote(allRemotes) {
+        var defaultRemotes = Preferences.get("defaultRemotes") || {},
+            candidate = defaultRemotes[Utils.getProjectRoot()];
+
+        var exists = _.find(allRemotes, function (remote) {
+            return remote.name === candidate;
+        });
+        if (!exists) {
+            candidate = null;
+            if (allRemotes.length > 0) {
+                candidate = _.first(allRemotes).name;
+            }
+        }
+
+        return candidate;
     }
 
     function setDefaultRemote(remoteName) {
@@ -72,7 +85,7 @@ define(function (require) {
     function refreshRemotesPicker() {
         Git.getRemotes().then(function (remotes) {
             // Set default remote name and cache the remotes dropdown menu
-            var defaultRemoteName    = getDefaultRemote();
+            var defaultRemoteName = getDefaultRemote(remotes);
 
             // Disable Git-push and Git-pull if there are not remotes defined
             $gitPanel